requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
lark-oapi==1.4.23
python-socks[asyncio]==2.7.2
APScheduler==3.10.4
cryptography==41.0.7
